fix(command): share claude transcript project resolver
