+++ src/agents/command/claude-cli-project-dir.ts
@@ -1,0 +1,45 @@
+import fs from "node:fs";
+import os from "node:os";
+import path from "node:path";
+import { normalizeOptionalString } from "../../shared/string-coerce.js";
+
+const CLAUDE_PROJECTS_DIRNAME = path.join(".claude", "projects");
+const MAX_SANITIZED_PROJECT_LENGTH = 200;
+
+function simpleHash36(input: string): string {
+  let hash = 0;
+  for (let index = 0; index < input.length; index += 1) {
+    hash = (hash * 31 + input.charCodeAt(index)) >>> 0;
+  }
+  return hash.toString(36);
+}
+
+function sanitizeClaudeCliProjectKey(workspaceDir: string): string {
+  const sanitized = workspaceDir.replace(/[^a-zA-Z0-9]/g, "-");
+  if (sanitized.length <= MAX_SANITIZED_PROJECT_LENGTH) {
+    return sanitized;
+  }
+  return `${sanitized.slice(0, MAX_SANITIZED_PROJECT_LENGTH)}-${simpleHash36(workspaceDir)}`;
+}
+
+function canonicalizeWorkspaceDir(workspaceDir: string): string {
+  const resolved = path.resolve(workspaceDir).normalize("NFC");
+  try {
+    return fs.realpathSync.native(resolved).normalize("NFC");
+  } catch {
+    return resolved;
+  }
+}
+
+export function resolveClaudeCliProjectDirForWorkspace(params: {
+  workspaceDir: string;
+  homeDir?: string;
+}): string {
+  const homeDir = normalizeOptionalString(params.homeDir) || process.env.HOME || os.homedir();
+  const canonicalWorkspaceDir = canonicalizeWorkspaceDir(params.workspaceDir);
+  return path.join(
+    homeDir,
+    CLAUDE_PROJECTS_DIRNAME,
+    sanitizeClaudeCliProjectKey(canonicalWorkspaceDir),
+  );
+}
